fix the dromdowmenu
