+++ src/components/CardsData.js
@@ -1,0 +1,38 @@
+import React from "react";
+
+export default function CardsData() {
+  return [
+    {
+      title: 'Word Games',
+      description: 'Multi games web app.',
+      link: 'https://tailwindcss.com/docs/gap',
+      imgA: './unlv-badges/css-badge.png',
+      imgB: './unlv-badges/html-badge.png',
+      imgC: './unlv-badges/final-badge.png',
+    },
+    {
+      title: 'Card 2 Title',
+      description: 'Description for card 2',
+      link: 'https://tailwindcss.com/docs/gap',
+      imgA: './unlv-badges/css-badge.png',
+      imgB: './unlv-badges/html-badge.png',
+      imgC: './unlv-badges/final-badge.png',
+    },
+    {
+      title: 'Card 3 Title',
+      description: 'Description for card 3',
+      link: 'https://tailwindcss.com/docs/gap',
+      imgA: './unlv-badges/css-badge.png',
+      imgB: './unlv-badges/html-badge.png',
+      imgC: './unlv-badges/final-badge.png',
+    },
+    {
+      title: 'Card 4 Title',
+      description: 'Description for card 4',
+      link: 'https://tailwindcss.com/docs/gap',
+      imgA: './unlv-badges/css-badge.png',
+      imgB: './unlv-badges/html-badge.png',
+      imgC: './unlv-badges/final-badge.png',
+    },
+  ]
+}
