+++ src/components/ProjectList.jsx
@@ -1,41 +1,8 @@
 import React from "react";
+import CardsData from "./CardsData";
 
 export default function ProjectList() {
-  const cardsData = [
-    {
-      title: 'Word Games',
-      description: 'Multi games web app.',
-      link: 'https://tailwindcss.com/docs/gap',
-      imgA: './unlv-badges/css-badge.png',
-      imgB: './unlv-badges/html-badge.png',
-      imgC: './unlv-badges/final-badge.png',
-    },
-    {
-      title: 'Card 2 Title',
-      description: 'Description for card 2',
-      link: 'https://tailwindcss.com/docs/gap',
-      imgA: './unlv-badges/css-badge.png',
-      imgB: './unlv-badges/html-badge.png',
-      imgC: './unlv-badges/final-badge.png',
-    },
-    {
-      title: 'Card 3 Title',
-      description: 'Description for card 3',
-      link: 'https://tailwindcss.com/docs/gap',
-      imgA: './unlv-badges/css-badge.png',
-      imgB: './unlv-badges/html-badge.png',
-      imgC: './unlv-badges/final-badge.png',
-    },
-    {
-      title: 'Card 4 Title',
-      description: 'Description for card 4',
-      link: 'https://tailwindcss.com/docs/gap',
-      imgA: './unlv-badges/css-badge.png',
-      imgB: './unlv-badges/html-badge.png',
-      imgC: './unlv-badges/final-badge.png',
-    },
-  ]
-
+  const cardsData = CardsData()
   return (
     <>
       <section className="flex flex-col w-full p-4 gap-4 rounded-xl">
